requirements/base.txt
Django==4.2.3
django-tailwind==3.6.0 
django-environ==0.10.0 
celery==5.3.1 
redis==5.0.0
django-crontab==0.7.1 
django-cors-headers==4.2.0
djangorestframework==3.14.0 
djoser==2.2.0 
# visual for autentication
